+++ src/components/navigation/BreadcrumbEllipsis/interfaces/BreadcrumbEllipsis.Route.interface.ts
