+++ src/components/navigation/BreadcrumbEllipsis/interfaces/BreadcrumbEllipsis.interface.ts
@@ -1,0 +1,7 @@
+import { Typos } from "../types/BreadcrumbEllipsis.Typos.type";
+import { IRoute } from "./BreadcrumbEllipsis.Route.interface";
+
+export interface IBreadcrumbEllipsisProps {
+  typo?: Typos[number];
+  routes: IRoute[];
+}
