--- src/components/navigation/BreadcrumbEllipsis/interfaces/BreadcrumbEllipsisProps.interface.ts
@@ -1,7 +1,0 @@
-import { Typos } from "../types/Typos.type";
-import { IRoute } from "./Route.interface";
-
-export interface IBreadcrumbEllipsisProps {
-  typo?: Typos[number];
-  routes: IRoute[];
-}
